pyproject.toml
[project]
name = "launcher"
version = "0.1.0"
description = "A python launcher"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "textual>=3.1.0",
    "toml>=0.10.2",
    "wetlands>=0.1.4",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
]
